Removed tests directory, rename directories with data_ to datatier_, changed platform_ to datatier_
