+++ datatier_local/DDL-InsertData/db_platform.ddl
