+++ datatier_local/DDL-InsertData/platform-configuration_details.sql
