+++ datatier_local/DDL-InsertData/platform-platform_operations.sql
