+++ main.py
@@ -13,14 +13,14 @@ from common.platform_settings import build_platform_variables
 from common.platform_settings import build_platform_config
 import common.error_audit_mgmt
 from common.error_audit_mgmt import process_auditerror_details
-import platform_data_actions
-from platform_data_actions import datatier
-from platform_data_actions import platform
+import datatier_actions
+from datatier_actions import datatier
+from datatier_actions import platform
 
 def main():
     # Set Platform Variables
     start_datetime = datetime.now()
-    local_database_path = os.getcwd() + os.sep + "platform_data_local" + os.sep
+    local_database_path = os.getcwd() + os.sep + "datatier_local" + os.sep
     platform_vars = build_platform_variables();
     # Pull in platform configuration settings from configuration database
     platform_settings = build_platform_config(platform_vars.local_database_path);
@@ -41,8 +41,8 @@ def main():
                                    error_desc="NA", processed_objectname="NA", audit_details="NA")
         # Do platform_operation activities
         if (platform_settings.platform_operation_name == "data_synthetic_generation"):
-            list_generated_dataneeds = platform.query_platformdata_general_activerecords(platform_vars=platform_vars,
-                    platform_settings= platform_settings, sql_connection=rdbms_connection, table_name="platform_datageneration_dataattributes")
+            list_generated_dataneeds = platform.query_platformdata_general_activerecords(platform_vars = platform_vars,
+                    platform_settings = platform_settings, sql_connection=rdbms_connection, table_name="platform_datageneration_dataattributes")
             print("")
         elif(platform_settings.platform_operation_name == ""):
             print("No Operation Defined")
