--- settings/settings.ini
@@ -1,29 +1,0 @@
-[General]
-# Supported Values - True or False
-output_settings = False
-# Supported values - data-anon, data-deidentification, data-generation, data-retrieval, tokenization, tagging
-platform_operation = data-generation
-# Supported values - rdbms-postgres,rdbms-sqlserver, mongo, file,azure
-output_destination = rdbms-sqlserver
-
-[Local]
-local_storage_location = C:\\Data\\TestData\\
-local_extension = hl7
-
-[RDBMS_DataTier]
-# none, postgresql, sqlserver
-type = sqlserver
-host= localhost
-port_nmbr = 1234
-database = school
-
-[Connnectors_Details]
-# This is where the connection_name in the connectors section must be defined. ONLY the key details meeting
-# the specific formats, there SHOULD be no username or passwords. Those are defined in the specifically defined
-# .env file nd referenced and looked up by matching the connector_name between these two files.
-
-[Connectors]
-# Supported format is connector_type:connector_name. If there are multiple values then use a single space
-# Supported connector_types are files, redis
-connector_type=
-timing =
