--- tests/api_tester.py
@@ -1,11 +1,0 @@
-from datetime import datetime
-from unittest import TestCase
-
-# Main Program
-def main():
-    print(f"Start Type: {datetime.now}")
-
-    print("Program Ended")
-
-if __name__ == "__main__":
-    main()
